+++ eelbrain/tests/test_table.py
@@ -45,8 +45,8 @@ def test_difference():
     dds = table.difference('fltvar', 'A', 'a1', 'a2', 'rm', ds=ds)
     assert repr(dds) == "<Dataset n_cases=20 {'rm':F, 'fltvar':V, 'rmf':F}>"
     assert_array_equal(dds['rmf'], Factor(dds['rm'], labels=labels))
-    dds = table.difference('fltvar', 'A', 'a1', 'a2', 'rm', by='B', ds=ds)
-    assert repr(dds) == "<Dataset n_cases=40 {'rm':F, 'fltvar':V, 'B':F, 'rmf':F}>"
+    dds = table.difference('fltvar', 'A', 'a1', 'a2', 'rm % B', ds=ds)
+    assert repr(dds) == "<Dataset n_cases=40 {'rm':F, 'B':F, 'fltvar':V, 'rmf':F}>"
     # difference of the difference
     ddds = table.difference('fltvar', 'B', 'b1', 'b2', 'rm', ds=dds)
     assert repr(ddds) == "<Dataset n_cases=20 {'rm':F, 'fltvar':V, 'rmf':F}>"
@@ -58,10 +58,10 @@ def test_difference():
     ds['C', :] = 'c1'
     ds2['C', :] = 'c2'
     ds = combine((ds, ds2))
-    dds = table.difference('fltvar', 'A', 'a1', 'a2', 'rm', by='B%C', ds=ds)
-    assert repr(dds) == "<Dataset n_cases=80 {'rm':F, 'fltvar':V, 'B':F, 'C':F, 'rmf':F}>"
-    dds = table.difference('fltvar', 'A%B', ('a1', 'b1'), ('a2', 'b2'), 'rm', by='C', ds=ds)
-    assert repr(dds) == "<Dataset n_cases=40 {'rm':F, 'fltvar':V, 'C':F, 'rmf':F}>"
+    dds = table.difference('fltvar', 'A', 'a1', 'a2', 'rm % B % C', ds=ds)
+    assert repr(dds) == "<Dataset n_cases=80 {'rm':F, 'B':F, 'C':F, 'fltvar':V, 'rmf':F}>"
+    dds = table.difference('fltvar', 'A%B', ('a1', 'b1'), ('a2', 'b2'), 'rm % C', ds=ds)
+    assert repr(dds) == "<Dataset n_cases=40 {'rm':F, 'C':F, 'fltvar':V, 'rmf':F}>"
 
 
 def test_frequencies():